fix: add explicit casts for 64-to-32 bit truncation [-Wshorten-64-to-32]
On 64-bit platforms (macOS/Linux), several arithmetic expressions produce
long long results that were implicitly truncated to unsigned/int:

- sndchip.cpp: end_chip_tick computation via long long arithmetic
- sndrender.cpp: endtick computation via long long arithmetic
- sid.cpp: bit() helper returned long; return type changed to unsigned int
- soloud_audiosource.cpp: seek() passes long samples to unsigned int params
- soloud_thread.cpp: tv_sec (time_t, 64-bit) multiplied then returned as int

+++ src/soloud/include/zx7decompress.h
@@ -78,7 +78,7 @@ static void zx7_write_byte(struct zx7_io *io, int value) {
 static void zx7_write_bytes(struct zx7_io *io, int offset, int length) {
     int i;
     while (length-- > 0) {
-        i = io->output_index - offset;
+        i = (int)io->output_index - offset;
         zx7_write_byte(io, io->output_data[i]);
     }
 }
@@ -101,7 +101,7 @@ static int zx7_decompress(unsigned char *input_data, unsigned char *output_data)
         } else {
             length = zx7_read_elias_gamma(&io) + 1;
             if (length == 0) {
-                return io.input_index;
+                return (int)io.input_index;
             }
             zx7_write_bytes(&io, zx7_read_offset(&io) + 1, length);
         }
